requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
alpaca-trade-api==3.2.0
annotated-types==0.6.0
appdirs==1.4.4
astunparse==1.6.3
attrs==23.2.0
backtrader==1.9.78.123
beautifulsoup4==4.12.3
black==24.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
deprecation==2.1.0
flatbuffers==24.3.7
fonttools==4.50.0
frozendict==2.4.0
frozenlist==1.4.1
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
html5lib==1.1
idna==3.6
iniconfig==2.0.0
joblib==1.3.2
keras==3.1.1
kiwisolver==1.4.5
libclang==18.1.1
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.3.2
msgpack==1.0.3
multidict==6.0.5
multitasking==0.0.11
mypy-extensions==1.0.0
namex==0.0.7
numpy==1.26.4
opt-einsum==3.3.0
optree==0.10.0
packaging==24.0
pandas==2.2.1
pathspec==0.12.1
patsy==0.5.6
peewee==3.17.1
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
protobuf==4.25.3
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pyparsing==3.1.2
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
QuantLib==1.33
requests==2.31.0
rich==13.7.1
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
setuptools==69.2.0
six==1.16.0
soupsieve==2.5
sseclient-py==1.8.0
statsmodels==0.14.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
threadpoolctl==3.4.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.18
webencodings==0.5.1
websocket-client==1.7.0
websockets==10.4
Werkzeug==3.0.1
wheel==0.43.0
wrapt==1.16.0
yarl==1.9.4
yfinance==0.2.37

plotly~=5.20.0
dash~=2.16.1
Flask~=3.0.3
eventlet~=0.30.2
gevent~=24.2.1